+++ jkube-kit/resource/helm/src/main/java/org/eclipse/jkube/kit/resource/helm/HelmService.java
@@ -22,11 +22,13 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
+import java.util.function.BiPredicate;
 import java.util.function.Consumer;
 import java.util.function.UnaryOperator;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import io.fabric8.kubernetes.client.utils.Serialization;
 import org.eclipse.jkube.kit.common.JKubeConfiguration;
 import org.eclipse.jkube.kit.common.KitLogger;
 import org.eclipse.jkube.kit.common.RegistryConfig;
@@ -49,7 +51,10 @@ import org.apache.commons.compress.archivers.tar.TarArchiveEntry;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang3.StringUtils;
 
+
 import static org.apache.commons.lang3.StringUtils.EMPTY;
+import static org.eclipse.jkube.kit.common.JKubeFileInterpolator.DEFAULT_FILTER;
+import static org.eclipse.jkube.kit.common.JKubeFileInterpolator.interpolate;
 import static org.eclipse.jkube.kit.common.util.MapUtil.getNestedMap;
 import static org.eclipse.jkube.kit.common.util.TemplateUtil.escapeYamlTemplate;
 import static org.eclipse.jkube.kit.resource.helm.HelmServiceUtil.isRepositoryValid;
@@ -60,6 +65,7 @@ public class HelmService {
 
   private static final String YAML_EXTENSION = ".yaml";
   private static final String CHART_API_VERSION = "v1";
+  private static final String FRAGMENT_FILE_EXTENSION = ".helm.yaml";
   private static final String CHART_FILENAME = "Chart" + YAML_EXTENSION;
   private static final String VALUES_FILENAME = "values" + YAML_EXTENSION;
   private static final String GOLANG_EXPRESSION_REGEX = "\\{\\{.+}}";
@@ -94,7 +100,7 @@ public class HelmService {
       logger.debug("Processing source files");
       processSourceFiles(sourceDir, templatesDir);
       logger.debug("Creating %s", CHART_FILENAME);
-      createChartYaml(helmConfig, outputDir);
+      createChartYaml(helmConfig, outputDir, jKubeConfiguration);
       logger.debug("Copying additional files");
       copyAdditionalFiles(helmConfig, outputDir);
       logger.debug("Gathering parameters for placeholders");
@@ -230,8 +236,65 @@ public class HelmService {
     }
   }
 
-  static void createChartYaml(HelmConfig helmConfig, File outputDir) throws IOException {
-    final Chart chart = new Chart();
+  static void createChartYaml(HelmConfig helmConfig, File outputDir, JKubeConfiguration jKubeConfiguration) throws IOException {
+    final Chart chartFromHelmConfig = createChartFromHelmConfig(helmConfig);
+    final Chart chartFromFragment = createChartFromFragment(jKubeConfiguration);
+    final Chart mergedChart = mergeCharts(chartFromHelmConfig, chartFromFragment);
+
+    File outputChartFile = new File(outputDir, CHART_FILENAME);
+    ResourceUtil.save(outputChartFile, mergedChart, ResourceFileType.yaml);
+  }
+
+  private static Chart mergeCharts(Chart chartFromHelmConfig, Chart chartFromFragment) {
+    if (chartFromFragment == null) {
+      return chartFromHelmConfig;
+    }
+    Chart.ChartBuilder chartBuilder = chartFromFragment.toBuilder();
+    updateOriginal((c1, c2) -> StringUtils.isBlank(c1.getApiVersion()) && StringUtils.isNotBlank(c2.getApiVersion()),
+        chartBuilder, chartFromFragment, chartFromHelmConfig,
+        cb -> cb.apiVersion(chartFromHelmConfig.getApiVersion()));
+    updateOriginal((c1, c2) -> StringUtils.isBlank(c1.getName()) && StringUtils.isNotBlank(c2.getName()),
+        chartBuilder, chartFromFragment, chartFromHelmConfig,
+        cb -> cb.name(chartFromHelmConfig.getName()));
+    updateOriginal((c1, c2) -> StringUtils.isBlank(c1.getHome()) && StringUtils.isNotBlank(c2.getHome()),
+        chartBuilder, chartFromFragment, chartFromHelmConfig,
+        cb -> cb.home(chartFromHelmConfig.getHome()));
+    updateOriginal((c1, c2) -> StringUtils.isBlank(c1.getVersion()) && StringUtils.isNotBlank(c2.getVersion()),
+        chartBuilder, chartFromFragment, chartFromHelmConfig,
+        cb -> cb.version(chartFromHelmConfig.getVersion()));
+    updateOriginal((c1, c2) -> StringUtils.isBlank(c1.getDescription()) && StringUtils.isNotBlank(c2.getDescription()),
+        chartBuilder, chartFromFragment, chartFromHelmConfig,
+        cb -> cb.description(chartFromHelmConfig.getDescription()));
+    updateOriginal((c1, c2) -> StringUtils.isBlank(c1.getEngine()) && StringUtils.isNotBlank(c2.getEngine()),
+        chartBuilder, chartFromFragment, chartFromHelmConfig,
+        cb -> cb.engine(chartFromHelmConfig.getEngine()));
+    updateOriginal((c1, c2) -> StringUtils.isBlank(c1.getIcon()) && StringUtils.isNotBlank(c2.getIcon()),
+        chartBuilder, chartFromFragment, chartFromHelmConfig,
+        cb -> cb.icon(chartFromHelmConfig.getIcon()));
+    updateOriginal((c1, c2) -> c1.getSources() == null && c2.getSources() != null,
+        chartBuilder, chartFromFragment, chartFromHelmConfig,
+        cb -> cb.sources(chartFromHelmConfig.getSources()));
+    updateOriginal((c1, c2) -> c1.getKeywords() == null && c2.getKeywords() != null,
+        chartBuilder, chartFromFragment, chartFromHelmConfig,
+        cb -> cb.keywords(chartFromHelmConfig.getKeywords()));
+    updateOriginal((c1, c2) -> c1.getMaintainers() == null && c2.getMaintainers() != null,
+        chartBuilder, chartFromFragment, chartFromHelmConfig,
+        cb -> cb.maintainers(chartFromHelmConfig.getMaintainers()));
+    updateOriginal((c1, c2) -> c1.getDependencies() == null && c2.getDependencies() != null,
+        chartBuilder, chartFromFragment, chartFromHelmConfig,
+        cb -> cb.dependencies(chartFromHelmConfig.getDependencies()));
+
+    return chartBuilder.build();
+  }
+
+  private static void updateOriginal(BiPredicate<Chart, Chart> condition, Chart.ChartBuilder chartBuilder, Chart original, Chart opinionated, Consumer<Chart.ChartBuilder> chartBuilderConsumer) {
+    if (condition.test(original, opinionated)) {
+      chartBuilderConsumer.accept(chartBuilder);
+    }
+  }
+
+  private static Chart createChartFromHelmConfig(HelmConfig helmConfig) {
+    Chart chart = new Chart();
     chart.setApiVersion(CHART_API_VERSION);
     chart.setName(helmConfig.getChart());
     chart.setVersion(helmConfig.getVersion());
@@ -243,9 +306,34 @@ public class HelmService {
     chart.setKeywords(helmConfig.getKeywords());
     chart.setEngine(helmConfig.getEngine());
     chart.setDependencies(helmConfig.getDependencies());
+    return chart;
+  }
 
-    File outputChartFile = new File(outputDir, CHART_FILENAME);
-    ResourceUtil.save(outputChartFile, chart, ResourceFileType.yaml);
+  private static Chart createChartFromFragment(JKubeConfiguration jKubeConfiguration) {
+    File helmChartFragment = getChartYamlFileFromFragmentsDir(jKubeConfiguration.getResolvedResourceSourceDirs());
+    if (helmChartFragment != null && helmChartFragment.exists()) {
+      try {
+        String interpolatedFragmentContent = interpolate(helmChartFragment, jKubeConfiguration.getProperties(), DEFAULT_FILTER);
+        return Serialization.yamlMapper().readValue(interpolatedFragmentContent, Chart.class);
+      } catch (IOException e) {
+        throw new IllegalArgumentException("Failure in parsing Helm Chart fragment : " + e.getMessage());
+      }
+    }
+    return null;
+  }
+
+  private static File getChartYamlFileFromFragmentsDir(List<File> fragmentDirs) {
+    File foundChartYamlFile = null;
+    if (fragmentDirs != null) {
+      Optional<File> fileOptional = fragmentDirs.stream()
+          .map(f -> new File(f, String.format("%s%s", "Chart", FRAGMENT_FILE_EXTENSION)))
+          .filter(File::exists)
+          .findFirst();
+      if (fileOptional.isPresent()) {
+        foundChartYamlFile = fileOptional.get();
+      }
+    }
+    return foundChartYamlFile;
   }
 
   private static void copyAdditionalFiles(HelmConfig helmConfig, File outputDir) throws IOException {