+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Usage:
 ```
 ### 1.13-SNAPSHOT
 * Fix #1478: Should detect and warn the user if creating ingress and ingress controller not available
+* Fix #2092: Support for `Chart.yaml` fragments
 * Fix #2150: Bump Kubernetes Client to 6.6.0 (fixes issues when trace-logging OpenShift builds)
 * Fix #2162: Bump Kubernetes Client to 6.6.1 (HttpClient with support for PUT + InputStream)
 * Fix #2165: Introduce a Kubernetes resource Security Hardening profile (opt-in)